server/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
lxml==4.6.3
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.6.2
Pillow==8.2.0
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.60.0
urllib3==1.26.4
Werkzeug==1.0.1
